+++ NEWS.md
@@ -1,3 +1,7 @@
+MixedModels v3.9.0 Release Notes
+========================
+* Add support for `StatsModels.formula` [#536]
+
 MixedModels v3.8.0 Release Notes
 ========================
 * Add support for NLopt `maxtime` option to `OptSummary` [#524]
@@ -204,3 +208,4 @@ Package dependencies
 [#518]: https://github.com/JuliaStats/MixedModels.jl/issues/518
 [#523]: https://github.com/JuliaStats/MixedModels.jl/issues/523
 [#524]: https://github.com/JuliaStats/MixedModels.jl/issues/524
+[#536]: https://github.com/JuliaStats/MixedModels.jl/issues/536